requirements.txt
streamlit
pandas
biopython
python-dotenv
google-genai
requests
lxml
